+++ index.ts
@@ -4,7 +4,7 @@
 
 import messaging from '@react-native-firebase/messaging';
 import { AppRegistry } from 'react-native';
-import App from './src/pages/App';
+import App from './src/App';
 import { name as appName } from './app.json';
 
 messaging().setBackgroundMessageHandler(async (remoteMessage) => {
