--- src/pages/App.tsx
@@ -1,48 +1,0 @@
-import messaging from '@react-native-firebase/messaging';
-import { NavigationContainer } from '@react-navigation/native';
-import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import { useEffect } from 'react';
-import { Alert } from 'react-native';
-import { SafeAreaProvider } from 'react-native-safe-area-context';
-import HomeScreen from './HomeScreen';
-
-const Stack = createNativeStackNavigator();
-
-const App = () => {
-  const requestUserPermission = async () => {
-    const authStatus = await messaging().requestPermission();
-    const enable =
-      authStatus === messaging.AuthorizationStatus.AUTHORIZED ||
-      authStatus === messaging.AuthorizationStatus.PROVISIONAL;
-
-    if (enable) {
-      console.log('Authorization status: ', authStatus); //1: access
-    }
-  };
-
-  useEffect(() => {
-    requestUserPermission();
-  }, []);
-
-  //Foreground
-  useEffect(() => {
-    const unsubscribe = messaging().onMessage(async (remoteMessage) => {
-      console.log('>> A new FCM message arrived!');
-      Alert.alert('A new FCM message arrived!', JSON.stringify(remoteMessage));
-    });
-
-    return unsubscribe;
-  }, []);
-
-  return (
-    <NavigationContainer>
-      <SafeAreaProvider>
-        <Stack.Navigator initialRouteName="Home" screenOptions={{ headerShown: false }}>
-          <Stack.Screen name="Home" component={HomeScreen} />
-        </Stack.Navigator>
-      </SafeAreaProvider>
-    </NavigationContainer>
-  );
-};
-
-export default App;
